Control works!

+++ test-19/Core/Inc/MotorControl.h
@@ -46,7 +46,7 @@ public:
 	// For between methods
 	float last_error = 0;
     float error = 0;
-    int16_t running_error = 0; // You don't want this to be a float for integral windup problems.
+    float running_error = 0;
     float error_slope;
     int16_t effort = 0;
     int16_t saturation_limit;
